refactored the code and fixed signup issue, it works now
