+++ models/User.js
@@ -28,7 +28,7 @@ User.init(
         twitter: {
             type: DataTypes.STRING,
             allowNull: true
-        },   
+        }, 
         email: {
           type: DataTypes.STRING,
           allowNull: false,
