+++ public/javascript/delete-post.js
@@ -1,26 +1,26 @@
 async function deleteFormHandler(event) {
-    event.preventDefault();
-    
-    const id = window.location.toString().split('/')[
-        window.location.toString().split('/').length - 1
-      ];
-
-    const response = await fetch(`/api/posts/${id}`, {
-        method: 'DELETE',
-        body: JSON.stringify({
-          post_id: id
-        }),
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      });
-      
-      if (response.ok) {
-        document.location.replace('/dashboard/');
-      } else {
-        alert(response.statusText);
-      }
-    
-  }
+  event.preventDefault();
   
-  document.querySelector('.delete-post-btn').addEventListener('click', deleteFormHandler);
+  const id = window.location.toString().split('/')[
+      window.location.toString().split('/').length - 1
+    ];
+
+  const response = await fetch(`/api/posts/${id}`, {
+      method: 'DELETE',
+      body: JSON.stringify({
+        post_id: id
+      }),
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    });
+    
+    if (response.ok) {
+      document.location.replace('/dashboard/');
+    } else {
+      alert(response.statusText);
+    }
+  
+}
+
+document.querySelector('.delete-post-btn').addEventListener('click', deleteFormHandler);
