+++ public/javascript/logout.js
@@ -1,14 +1,14 @@
 async function logout() {
-    const response = await fetch('/api/users/logout', {
-      method: 'post',
-      headers: { 'Content-Type': 'application/json' }
-    });
-  
-    if (response.ok) {
-      document.location.replace('/');
-    } else {
-      alert(response.statusText);
-    }
+  const response = await fetch('/api/users/logout', {
+    method: 'post',
+    headers: { 'Content-Type': 'application/json' }
+  });
+
+  if (response.ok) {
+    document.location.replace('/');
+  } else {
+    alert(response.statusText);
   }
-  
-  document.querySelector('#logout').addEventListener('click', logout);
+}
+
+document.querySelector('#logout').addEventListener('click', logout);
