+++ utils/auth.js
@@ -1,9 +1,9 @@
 const withAuth = (req, res, next) => {
-    if (!req.session.user_id) {
-      res.redirect('/login');
-    } else {
-      next();
-    }
-  };
-  
-  module.exports = withAuth;
+  if (!req.session.user_id) {
+    res.redirect('/login');
+  } else {
+    next();
+  }
+};
+
+module.exports = withAuth;
