+++ views/create-post.handlebars
@@ -15,4 +15,4 @@
   </form>
 </section>
 
-<script src="/javascript/add-post.js"></script>
+<script src="/javascript/add-post.js"></script>
