+++ views/dashboard.handlebars
@@ -1,13 +1,13 @@
     <div class="mb-5 pt-3">
         <a href="/dashboard/create/"><button id="create-post" class="button is-primary is-pulled-right">create post</button></a>
-        
+
         <h2 class="title pb-4">Your posts</h2><hr>
 {{#if posts.length}}
     </div>
 
     {{#each posts as |post|}}
         {{> post-info post}}
-        
+
             <a href="/dashboard/edit/{{post.id}}"><button class="button mb-5 ml-5 is-warning">edit post</button></a>
         </div>
     {{/each}}
@@ -16,5 +16,3 @@
 <h6 class="pb-4 has-text-centered is-italic">You do not currently have posts</h6>
 
 {{/if}}
-
-
