+++ views/edit-post.handlebars
@@ -24,4 +24,4 @@
 
 <script src="/javascript/edit-post.js"></script>
 <script src="/javascript/delete-post.js"></script>
-<script src="/javascript/comment.js"></script>
+<script src="/javascript/comment.js"></script>
