+++ views/layouts/main.handlebars
@@ -12,8 +12,7 @@
 
   <div class="wrapper">
     <header>
-      <img src="/public/assets/crypto.png" width="220px"  alt="search engine optimization" class="float-left" />
-        {{!-- <h1 class="has-text-black px-3 py-3 has-text-weight-bold">Tech Buzzing</h1> --}}
+      <img src="/assets/crypto.png" width="250px" alt="search engine optimization" class="float-left" />
         <nav class="navbar mt-2">
             {{#if loggedIn}}
              <input type="text" placeholder="Search for Crypto...">
@@ -28,14 +27,10 @@
         </nav>
     </header>
 
-    {{#if loggedIn}}
     <main>
             {{{ body }}}
     </main>
-    {{else}}
 
-
-    {{/if}}
 
     <footer class="footer rows pb-2">
         <div class="row-is-full content has-text-centered">
@@ -50,8 +45,8 @@
         </div>
       <div class="row-is-full content has-text-centered">
     <p>
-      <strong>Tech Buzzing</strong> by <a href="https://github.com/valerrianna/cryptobuzz.git">Randolph Richard , Valerie Nguyen, Skyler Alexander</a> | made with <a href="https://bulma.io/">bulma</a> | source code:
-      <a href="https://github.com/RandolphRichard/MVC-Techs-Blog.git"><i class="fab fa-github"></i></a> |  
+      <strong>CryptoBuzz</strong> by <a href="https://github.com/valerrianna/cryptobuzz.git">Randolph Richard , Valerie Nguyen, Skyler Alexander</a> | made with <a href="https://bulma.io/">bulma</a> | source code:
+      <a href="https://github.com/valerrianna/cryptobuzz.git><i class="fab fa-github"></i></a> |  
     </p>
   </div>
     </footer>
@@ -62,4 +57,4 @@
 {{/if}}
 </body>
 
-</html>
+</html>
