+++ views/login.handlebars
@@ -17,4 +17,4 @@
 </div>
 
 <!-- javascript file-->
-<script src="/javascript/login.js"></script>
+<script src="/javascript/login.js"></script>
