+++ views/partials/comments.handlebars
@@ -9,4 +9,4 @@
     </div>
   </section>
   {{/each}}
-</div>
+</div>
