+++ views/partials/post-info.handlebars
@@ -1,6 +1,6 @@
 <div class="card mb-6">
   <header class="card-header card-header-title has-background-grey-lighter">
-    
+
         <div class="column">
             <p class=" has-text-primary-light"></p>
       <a href="/post/{{id}}"  class="has-text-black is-capitalized is-pulled-left">{{title}} </a></p>
@@ -9,7 +9,7 @@
       <div class="column">
       <span class=""> </span>
       </div>
-    
+
   </header>
 
     <div class="card-content">
@@ -22,5 +22,5 @@
                 <a href="https://github.com/{{user.github}}"><i class="fab fa-github"></i></a>
                  | {{format_date created_at}} | <a href="/post/{{id}}">{{comments.length}} {{format_plural "comment" comments.length}}</a>
             </div>
-        
+
     </div>
