+++ views/signup.handlebars
@@ -24,4 +24,4 @@
 </div>
 
 <!-- javascript file-->
-<script src="/javascript/signup.js"></script>
+<script src="/javascript/signup.js"></script>
