+++ views/single-post.handlebars
@@ -12,7 +12,7 @@
   <div>
     <button class="button is-warning ml-5 mb-5" type="submit">Add your comment</button>
   </div>
-  
+
 </form>
 {{/if}}
 
@@ -20,4 +20,4 @@
 
 {{#if loggedIn}}
 <script src="/javascript/comment.js"></script>
-{{/if}}
+{{/if}}
